Stop 500's on a non-JSON /auth request
When the request is not JSON (and the `force` flag is False), Flask's
`get_json()` will return None:
https://github.com/mitsuhiko/flask/blob/0.10.1/flask/wrappers.py#L127

If somebody POSTs to `/auth` with a non-JSON mimetype, the server will
500 with `'NoneType' object has no attribute 'get'`.

Additionally, the default second parameter to `get()` is already None -
it can be safely omitted.

And more than one criterion are criteria. =)